chore: Now using jotai for modal state
