+++ components/Layout.tsx
@@ -1,15 +1,17 @@
 'use client'
 
-import { Modal, useModalContext } from 'contexts/ModalContext'
-
 import AccountModal from 'components/Modals/Account'
 import SignInModal from 'components/Modals/SignIn'
 import SignUpModal from 'components/Modals/SignUp'
+import { useAtom } from 'jotai'
 import { useEffect } from 'react'
 import { useAuthFunctions } from 'state/AuthFunctions'
+import { useModalFunctions } from 'state/ModalFunctions'
+import { Modal, modalsAtom } from 'state/state'
 
 export default function Layout({ children }: { children?: React.ReactNode }) {
-  const { modals, close } = useModalContext()
+  const [modals] = useAtom(modalsAtom)
+  const { close } = useModalFunctions()
   const { check } = useAuthFunctions()
 
   // Check if the user is authenticated
